requirements/common.txt
django
django-rest-framework
drf_yasg
djangorestframework-simplejwt